agent-trading/requirements.txt
# agent-trading dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
py-clob-client>=0.17.0
eth-account>=0.11.0
web3>=6.0.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=1.0.0
